Create new JS file
